LATAM_remote/InfoVac-main/requirements.txt
alabaster==0.7.12
appdirs==1.4.4
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.3
blis==0.4.1
boto3==1.16.14
botocore==1.19.14
bs4==0.0.1
catalogue==2.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
courlan==0.2.2
coverage==5.3
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.4
decorator==4.4.2
docutils==0.16
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.0.0a0/es_core_news_md-3.0.0a0.tar.gz
fake-useragent==0.1.11
fastapi==0.61.2
flake8==3.8.4
h11==0.11.0
htmldate==0.7.1
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jusText==2.2.0
kiwisolver==1.3.1
lime==0.2.0.1
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
murmurhash==1.0.4
networkx==2.5
nltk==3.5
numpy==1.19.2
packaging==20.4
pandas==1.1.3
parse==1.18.0
parso==0.7.1
pathy==0.3.1
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
prompt-toolkit==3.0.7
pycodestyle==2.6.0
pydantic==1.6.1
pyee==7.0.4
pyflakes==2.2.0
Pygments==2.7.1
pyparsing==2.4.7
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytokenizations==0.7.2
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pyzmq==19.0.2
readability-lxml==0.8.1
regex==2020.10.28
requests==2.24.0
requests-file==1.5.1
requests-html==0.10.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
smart-open==2.2.1
snowballstemmer==2.0.0
soupsieve==2.0.1
spacy==2.3.2
spacy-nightly==3.0.0rc2
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
srsly==2.3.1
starlette==0.13.6
thinc==8.0.0rc2
threadpoolctl==2.1.0
tifffile==2020.10.1
tldextract==2.2.3
tornado==6.0.4
tqdm==4.50.2
trafilatura==0.5.2
traitlets==5.0.4
typer==0.3.2
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.10
uvicorn==0.12.2
w3lib==1.22.0
wasabi==0.8.0
wcwidth==0.2.5
websockets==8.1
zipp==3.3.0
